AcademicMCP/requirements.txt

fastapi
uvicorn[standard]
google-generativeai
requests
scholarly
pymupdf
python-dotenv